Add manifest localization functionality (#1889)
This adds support for localization files, which Snaps can use to provide
translations for the Snap manifest.

Snaps can now specify a `locales` field in the `source` object, e.g.:

```json
{
  "source": {
    "locales": [
      "locales/en.json",
      "locales/nl.json"
    ]
  }
}
```

It consists of (relative) paths to locale files. These files contain the
translations which can be used in the snap manifest. For example:

```json
{
  "locale": "en",
  "messages": {
    "name": {
      "message": "Localization Example Snap",
      "description": "The name of the snap, which is used in the Snap manifest."
    },
    "description": {
      "message": "MetaMask example snap demonstrating how to localize a Snap.",
      "description": "The description of the snap, which is used in the Snap manifest."
    }
  }
}
```

Which can be used in the manifest as follows:

```json
{
  "proposedName": "{{ name }}",
  "description": "{{ description }}"
}
```