feat: support additional components inside forms (#2497)
Adds support for additional components inside forms, essentially
allowing all children of `Box` to also be children of `Form`. Although
`Form` remains the only component where `Field` can exist. **This also
means forms can now be nested, when this happens the input state is tied
to the nearest parent form.**

To support this change, a refactor of the `constructState` function was
done. This function now uses `walkJsx` to walk the entire JSX tree and
collect state values.

Closes https://github.com/MetaMask/snaps/issues/2496