Support conditional children in most JSX components (#2506)
This adds support for conditional children in most JSX components. The
exceptions are `Field`, `Row`, and other components which don't have
children in the first place. For example, the following is now possible:

```tsx
<Box direction="horizontal" alignment="space-between">
  {condition && <Text>Hello</Text>}
</Box>
```

To support this, I've added `boolean` as valid prop for the JSX
components. The `getJsxChildren` function filters out any falsy or
literal `true` values.

I've also added a small contributing guide for conventions around JSX
components, including the support for conditional children.

Closes #2505.

---------

Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>